Move MshSegment out to examples/demo.rs
Strongly-typed message header segment does not conform with generic
segment semantics, nor does it serve a specific purpose in the lib.

Remove the MshSegment struct and implementations from the main
library sources, place into the `demo` example to show conversion
semantics and accessor mechanism.
